# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pro Se Besties API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type IntakeSubmissionIssueType =
  (typeof IntakeSubmissionIssueType)[keyof typeof IntakeSubmissionIssueType];

export const IntakeSubmissionIssueType = {
  discrimination: "discrimination",
  harassment: "harassment",
  retaliation: "retaliation",
  wrongful_termination: "wrongful_termination",
  hostile_work_environment: "hostile_work_environment",
  wage_theft: "wage_theft",
  other: "other",
} as const;

export interface IntakeSubmission {
  /** User's description of the incident */
  what_happened: string;
  issue_type: IntakeSubmissionIssueType;
  /** US state abbreviation (e.g. CA, TX) */
  state: string;
  zip_code: string;
  employer_name: string;
  employer_location: string;
  /** Date the incident occurred (YYYY-MM-DD) */
  incident_date: string;
  /** Last day of employment if terminated (YYYY-MM-DD) */
  last_day_of_employment?: string | null;
  /** Whether the user reported the issue internally */
  reported_internally: boolean;
  /** Whether the user has supporting documentation */
  has_documentation: boolean;
  additional_context?: string | null;
}

export type NextStepPriority =
  (typeof NextStepPriority)[keyof typeof NextStepPriority];

export const NextStepPriority = {
  immediate: "immediate",
  short_term: "short_term",
  long_term: "long_term",
} as const;

export interface NextStep {
  title: string;
  description: string;
  priority: NextStepPriority;
  link?: string | null;
}

export interface RelevantAgency {
  name: string;
  acronym: string;
  description: string;
  url: string;
  filing_deadline_note?: string | null;
}

export interface IntakeResult {
  id: number;
  submission: IntakeSubmission;
  /** AI-generated summary of the case (placeholder for now) */
  case_summary: string;
  /** Detected legal issue categories */
  issue_categories: string[];
  /** Educational next steps */
  next_steps: NextStep[];
  /** Questions to prepare when speaking with an attorney */
  attorney_questions: string[];
  /** Relevant government agencies or filing bodies */
  relevant_agencies: RelevantAgency[];
  created_at: string;
}

export interface InsightQuestion {
  question: string;
  category?: string | null;
  submitter_name?: string | null;
}

export interface InsightItem {
  id: number;
  question: string;
  answer?: string | null;
  category?: string | null;
  contributor_name?: string | null;
  contributor_title?: string | null;
  is_featured: boolean;
  created_at: string;
}

export type ReferralItemCategory =
  (typeof ReferralItemCategory)[keyof typeof ReferralItemCategory];

export const ReferralItemCategory = {
  lawyer: "lawyer",
  therapist: "therapist",
  consultant: "consultant",
  community: "community",
} as const;

export interface ReferralItem {
  id: number;
  name: string;
  category: ReferralItemCategory;
  description: string;
  location?: string | null;
  is_remote: boolean;
  website?: string | null;
  phone?: string | null;
  email?: string | null;
  specialties: string[];
  accepts_pro_se?: boolean | null;
  sliding_scale?: boolean | null;
  created_at: string;
}

export interface ReferralCategorySummary {
  category: string;
  count: number;
  description: string;
}

export type ListReferralsParams = {
  category?: ListReferralsCategory;
};

export type ListReferralsCategory =
  (typeof ListReferralsCategory)[keyof typeof ListReferralsCategory];

export const ListReferralsCategory = {
  lawyer: "lawyer",
  therapist: "therapist",
  consultant: "consultant",
  community: "community",
} as const;