# lib/api-zod/src/generated/types/intakeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pro Se Besties API specification
 * OpenAPI spec version: 0.1.0
 */
import type { IntakeSubmission } from "./intakeSubmission";
import type { NextStep } from "./nextStep";
import type { RelevantAgency } from "./relevantAgency";

export interface IntakeResult {
  id: number;
  submission: IntakeSubmission;
  /** AI-generated summary of the case (placeholder for now) */
  case_summary: string;
  /** Detected legal issue categories */
  issue_categories: string[];
  /** Educational next steps */
  next_steps: NextStep[];
  /** Questions to prepare when speaking with an attorney */
  attorney_questions: string[];
  /** Relevant government agencies or filing bodies */
  relevant_agencies: RelevantAgency[];
  created_at: string;
}