# lib/api-zod/src/generated/types/intakeSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pro Se Besties API specification
 * OpenAPI spec version: 0.1.0
 */
import type { IntakeSubmissionIssueType } from "./intakeSubmissionIssueType";

export interface IntakeSubmission {
  /** User's description of the incident */
  what_happened: string;
  issue_type: IntakeSubmissionIssueType;
  /** US state abbreviation (e.g. CA, TX) */
  state: string;
  zip_code: string;
  employer_name: string;
  employer_location: string;
  /** Date the incident occurred (YYYY-MM-DD) */
  incident_date: string;
  /** Last day of employment if terminated (YYYY-MM-DD) */
  last_day_of_employment?: string | null;
  /** Whether the user reported the issue internally */
  reported_internally: boolean;
  /** Whether the user has supporting documentation */
  has_documentation: boolean;
  additional_context?: string | null;
}