# lib/api-zod/src/generated/types/nextStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pro Se Besties API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NextStepPriority } from "./nextStepPriority";

export interface NextStep {
  title: string;
  description: string;
  priority: NextStepPriority;
  link?: string | null;
}