# lib/api-zod/src/generated/types/relevantAgency.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pro Se Besties API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RelevantAgency {
  name: string;
  acronym: string;
  description: string;
  url: string;
  filing_deadline_note?: string | null;
}